SK:
-ENH: updated praxix to compile with the ITK_USE_REVIEW_STATISTICS cmake variable. Now it doesn't compile without anymore. The results of pxstatisticsonimage are slightlty difference for the quantiles, but that's really a small difference. The histogram is exactly the same (on which the quantile computation is based). NB: the output file in which the histogram is stored has changed a bit. For the absolute frequency counts the numbers are written as integers (for example: 5), instead of floating point numbers (like 5.00000000).

NB: just as a note: the new histogram class of itk stores the histogram frequencies as integers. So, normalizing a histogram seems to make no sense anymore, although some new ITK code still has functions that normalize a histogram. In praxix this was not used, so I didn't check what is the result of that operation, but keep it in mind.