ENH: Major refactoring of helper functions
Everything was put in the itktools namespace. A ITKToolsImageProperties file
was created to store all of these functions. Most things were moved out of
CommandLineArgumentHelper and into ITKToolsHelpers. Everything except
the class definition was moved out of ITKToolsBase and into ITKToolsHelpers.

The practice case for using some of these functions was BinaryImageOperator.